обновление

+++ app/src/main/java/ru/mobnius/vote/data/manager/vote/Vote.java
@@ -1,10 +1,10 @@
 package ru.mobnius.vote.data.manager.vote;
 
 public class Vote {
-    public int answerId;
-    public int questionId;
+    public long answerId;
+    public long questionId;
 
-    public Vote(int question, int answer) {
+    public Vote(long question, long answer) {
         answerId = answer;
         questionId = question;
     }
